moves the styles to the container root
