+++ src/signals/incident-management/containers/IncidentSplitContainer/styled.js
